Add sketch of main program
This is like a sketch with most functions mocked out for now. It's
vibe-coded where I have focused on the data structures, like the config
we have for the system, static and dynamic instances, config support for
running containers, what convert requests should look like and be
processed etc. It's been a good way to hash out the rough shape of the
system.

The easiest way to run the system now is with a static device instance,
i.e. use containerlab or something to run crpd separately, then add that
as a static instance. This is what I've done for now, so the default
config in the system reflects this.

I think the next step is to try and implement the convert function using
the netcli lib.