Flatten inner callback functions in convert()
Restructured the deeply nested callbacks in DeviceInstance.convert() to
instead define the callbacks on a single level. The callbacks still
form a chain of calls - one calls the other and so on. But now there is
less mental overhead when reading the code since the indentation level
is more consistent.

This is a pattern Kristian came up with in Orchestron (Device actor)
and I have grown to liking it too.