Refactor SchemaRegistry as fetch/compile coordinator
This is a PoC rewrite of the SchemaRegistry to act as a coordinator for
multiple devices fetching and compiling the same sets of schemas. On
startup, we want to avoid all devices (of the same type) fetching the
same schemas and then requesting compilation. Instead each
DeviceInstance actor only fetches the list of YANG modules (name,
revision) and creates a unique fingerprint (hash). It then uses this
fingerprint to register its intent to fetch and compile the schemas. The
first device to register does that, the others skip that step and
receive the compiled schemas via a callback. Schema compilation also
runs async in a SchemaCompiler actor, one instance per device per
module-set.